Install linter

+++ lib/tasks/delete_duplicate_technology_topic.rake
@@ -1,7 +1,5 @@
 desc "Delete the duplicated technology topic"
 task delete_duplicated_technology_topic: :environment do
   topic = Topic.find_by(id: 16, path: "/service-manual/technology")
-  if topic && topic.topic_sections.empty?
-    topic.destroy
-  end
+  topic.destroy if topic && topic.topic_sections.empty?
 end
